Create .py to .vhdl Script
A full script for .py to .vhdl conversion inlcuding 13 different applications to perform. It can be edited and re-arranged accordingly and it can be expanded further. 

Some codes (.vhdl) might need to be re-arranged and please check the pin output (.xdc) file before running the given codes to program a FPGA.